ADD: seeders, MODIFY: models and migrations
