+++ src/database/models/user.js
@@ -19,6 +19,8 @@ module.exports = (sequelize, DataTypes) => {
     firstName: DataTypes.STRING,
     lastName: DataTypes.STRING,
     email: DataTypes.STRING,
+    password: DataTypes.STRING,
+    address: DataTypes.STRING,
     avatar: DataTypes.STRING,
     roleId: DataTypes.INTEGER
   }, {
